refactor: remove all WASM artifacts, rename wasm_bundle → std_bundle
## Deleted
- pulsar_wasm_bundle crate (all files)
- generate_wasm_export() and helpers (rust_type_to_wasm, wasm_to_rust_param,
  rust_ret_to_wasm) from pulsar_macros — dead code since we dropped the
  wasm32 target
- wasm feature from pulsar_std Cargo.toml
- [target.'cfg(target_arch = "wasm32")'].dependencies getrandom entry
- All #[cfg(not(target_arch = "wasm32"))] gates from nodes/mod.rs —
  every module is always available on native targets

## Renamed
- crates/pulsar_wasm_bundle → crates/pulsar_std_bundle
  - package name: pulsar_std_bundle
  - description updated
  - src/lib.rs: PULSAR_STD_WASM → PULSAR_STD_LIB_BYTES (same const, cleaner name)
  - all test imports updated

## pulsar_macros cleanup
- wasm_export variable and #wasm_export in expanded quote removed
- wasm_safe attribute parameter renamed internally to native_only
- registry registration simplified (no longer two branches for wasm vs native)
- #[cfg(not(target_arch = "wasm32"))] retained only where needed by
  native_only nodes (those using std::thread etc.)

## pulsar_std
- nodes/mod.rs: removed all #[cfg(not(target_arch = "wasm32"))] gates;
  modules are unconditionally included (they're native-only by design)
- Cargo.toml: cdylib kept (needed for pulsar_std_bundle build), wasm feature gone
- registry.rs: comment updated

## Tests: 18 + 91 + 17 + 10 = 136 passing
