--- crates/pulsar_wasm_bundle/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "pulsar_wasm_bundle"
-version = "0.1.0"
-edition = "2021"
-description = "Pre-compiled pulsar_std WASM module embedded as bytes for the bytecode VM"
-build = "build.rs"
-
-[dependencies]
-
-[dev-dependencies]
-pbgc = { git = "https://github.com/Far-Beyond-Pulsar/PBGC.git", rev = "3605b51" }
-graphy = { git = "https://github.com/Far-Beyond-Pulsar/Graphy", rev = "74a993fe51efc00ac5e208af74a3df79f2c0ba47" }
-pulsar_std = { path = "../pulsar_std" }
-pulsar_bp_executor = { path = "../pulsar_bp_executor" }
-serde_json = "1.0"
-
-[build-dependencies]
